Refactor scoreHand and compareHands into ScoringUtils
